Package has to use underscores; consolidated tests;
