+++ .coveragerc
@@ -1,2 +1,2 @@
 [run]
-source = membersuite-api-client
+source = membersuite_api_client
