+++ .travis.yml
@@ -5,7 +5,8 @@ python:
 - '3.5'
 install:
 - pip install -r requirements.txt
-script: coverage run membersuite-api-client/tests.py
+- pip install -r requirements_test.txt
+script: nosetests --with-coverage
 after_success: coveralls
 notifications:
   hipchat: 5534a6204d6caa1a45ac2444282aca@WebDevActivity
