+++ requirements_test.txt
@@ -1,0 +1,3 @@
+coverage==4.2
+coveralls==1.1
+nose
